+++ ET-user/src/main/java/com/team1/etuser/user/dto/UserStocksRes.java
@@ -2,6 +2,8 @@ package com.team1.etuser.user.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
@@ -11,5 +13,6 @@ public class UserStocksRes {
     private String stockCode;
     private String stockImage;
     private String stockName;
-    private int amount;
+    private BigDecimal amount;
+    //보유종목
 }
